refactor(164): split game screen into packages
