+++ app/src/androidTest/java/org/cescfe/numpairs/feature/game/ui/GameScreenRobot.kt
@@ -26,6 +26,10 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.test.espresso.Espresso.pressBack as espressoPressBack
 import org.cescfe.numpairs.R
 import org.cescfe.numpairs.domain.puzzle.model.Operator
+import org.cescfe.numpairs.feature.game.ui.screen.GameScreenTestTags
+import org.cescfe.numpairs.feature.game.ui.semantics.GameHighlightedKey
+import org.cescfe.numpairs.feature.game.ui.semantics.OperandSelectorUsageHintVisualStateKey
+import org.cescfe.numpairs.feature.game.ui.semantics.StripEntryInputInvalidKey
 
 class GameScreenRobot(
     private val activity: ComponentActivity,
