+++ app/src/main/java/org/cescfe/numpairs/feature/game/ui/components/strip/AvailableNumberChip.kt
@@ -42,12 +42,12 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.cescfe.numpairs.domain.puzzle.model.Operator
-import org.cescfe.numpairs.feature.game.ui.OperandUsageIndicatorState
-import org.cescfe.numpairs.feature.game.ui.gameHighlightSemantics
-import org.cescfe.numpairs.feature.game.ui.operandUsageIndicatorColors
-import org.cescfe.numpairs.feature.game.ui.stripEntryInputInvalid
-import org.cescfe.numpairs.feature.game.ui.usageIndicatorContentDescriptionResId
-import org.cescfe.numpairs.feature.game.ui.usageIndicatorSymbol
+import org.cescfe.numpairs.feature.game.ui.indicators.OperandUsageIndicatorState
+import org.cescfe.numpairs.feature.game.ui.indicators.operandUsageIndicatorColors
+import org.cescfe.numpairs.feature.game.ui.indicators.usageIndicatorContentDescriptionResId
+import org.cescfe.numpairs.feature.game.ui.indicators.usageIndicatorSymbol
+import org.cescfe.numpairs.feature.game.ui.semantics.gameHighlightSemantics
+import org.cescfe.numpairs.feature.game.ui.semantics.stripEntryInputInvalid
 import org.cescfe.numpairs.ui.theme.NumPairsComponents
 import org.cescfe.numpairs.ui.theme.NumPairsTextStyles
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
