+++ app/src/main/java/org/cescfe/numpairs/feature/game/ui/screen/GameScreen.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game.ui
+package org.cescfe.numpairs.feature.game.ui.screen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -41,6 +41,8 @@ import org.cescfe.numpairs.feature.game.presentation.TileOperandOptionUiState
 import org.cescfe.numpairs.feature.game.presentation.TileOperandSelectionDialogUiState
 import org.cescfe.numpairs.feature.game.presentation.TileOperatorSelectionDialogUiState
 import org.cescfe.numpairs.feature.game.presentation.TileUiState
+import org.cescfe.numpairs.feature.game.ui.actions.RulesHelperAction
+import org.cescfe.numpairs.feature.game.ui.help.RulesHelperDialog
 import org.cescfe.numpairs.ui.theme.NumPairsComponents
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
