+++ app/src/main/java/org/cescfe/numpairs/feature/game/ui/screen/GameScreenSemantics.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game.ui
+package org.cescfe.numpairs.feature.game.ui.screen
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
